+++ index.html
@@ -229,7 +229,8 @@
       <button type="button" class="error__button">Попробовать снова</button>
     </div>
   </template>
-
-  <script src="js/main.js"></script>
+  <script type="module" src="js/main.js"></script>
+  <script type="module" src="js/util.js"></script>
+  <script type="module" src="js/data.js"></script>
 </body>
 </html>
